+++ src/test/java/com/bookspot/book/infra/search/BookOpenSearchRepositoryTest.java
@@ -1,6 +1,7 @@
 package com.bookspot.book.infra.search;
 
 import com.bookspot.book.infra.BookDocument;
+import com.bookspot.book.infra.BookSearchRepository;
 import com.bookspot.book.infra.search.cond.BookSearchCond;
 import com.bookspot.book.infra.search.result.BookPageResult;
 import org.springframework.beans.factory.annotation.Autowired;
